backend/pyproject.toml
[project]
name = "jobhunt-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.2",
    "fpdf2>=2.8.7",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.3",
    "pydantic-settings>=2.13.1",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.22",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.42.0",
]
